Добавила проверку на админа

+++ app/controllers/telegram/webhook_controller.rb
@@ -1,47 +1,68 @@
 module Telegram
   class WebhookController < Telegram::Bot::UpdatesController
-    def start!(*)
-      chat_id = TgChat.find_or_create_by(tg_id: chat['id'])
-      name = if chat['type'] == 'private'
-               chat['username']
-             else
-               chat['title']
-             end
-      chat_id.name = name
-      chat_id.status = true
-      chat_id.save!
+    def check_admin?
+      tg_id = from['id'].to_s
+      admin = '232193762'
 
-      respond_with :message, text: "Начинаю считать сообщения"
+      true if tg_id == admin
+    end
+
+    def start!(*)
+      if check_admin?
+        chat_id = TgChat.find_or_create_by(tg_id: chat['id'])
+        name = if chat['type'] == 'private'
+                 chat['username']
+               else
+                 chat['title']
+               end
+        chat_id.name = name
+        chat_id.status = true
+        chat_id.save!
+
+        respond_with :message, text: "Начинаю считать сообщения"
+      else
+        respond_with :message, text: "Вы не можете использовать эту команду"
+      end
     end
 
     def stats!(*)
-      count_message = TgMessage.where(tg_chat_id: chat['id']).count
-      respond_with :message, text: "В этом чате #{count_message} сообщений"
+      if check_admin?
+        count_message = TgMessage.where(tg_chat_id: chat['id']).count
+        respond_with :message, text: "В этом чате #{count_message} сообщений"
+      else
+        respond_with :message, text: "Вы не можете использовать эту команду"
+      end
     end
 
     def userstats!(*args)
-      respond_with :message, text: "Укажите имя пользователя" if args.empty?
+      if check_admin?
+        respond_with :message, text: "Укажите имя пользователя" if args.empty?
 
-      username = args[0].delete_prefix('@')
-      user = TgUser.find_by(tg_name: username)
+        username = args[0].delete_prefix('@')
+        user = TgUser.find_by(tg_name: username)
 
-      if user
-        user_message_count = TgMessage.where(tg_chat_id: chat['id'], tg_user_id: user.tg_id).count
-        respond_with :message, text: "#{user.name} отправил #{user_message_count} сообщений"
+        if user
+          user_message_count = TgMessage.where(tg_chat_id: chat['id'], tg_user_id: user.tg_id).count
+          respond_with :message, text: "#{user.name} отправил #{user_message_count} сообщений"
+        else
+          respond_with :message, text: "#{user.name} еще ничего не отправил"
+        end
       else
-        respond_with :message, text: "#{user.name} еще ничего не отправил"
+        respond_with :message, text: "Вы не можете использовать эту команду"
       end
     end
 
     def help!
-      respond_with :message,
-                   text: "/start запускает подсчет сообщений\n/stats показывает общее количество сообщений в чате\n/userstats @username показывает количество сообщний отдельного учатника чата"
+      if check_admin?
+        respond_with :message,
+                     text: "/start запускает подсчет сообщений\n/stats показывает общее количество сообщений в чате\n/userstats @username показывает количество сообщний отдельного учатника чата"
+      else
+        respond_with :message, text: "Вы не можете использовать эту команду"
+      end
     end
 
     def message(message)
       process_message(message)
-
-      # render plain: 'ok'
     end
 
     private
